bot/requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
flake8==5.0.4
greenlet==1.1.2
mccabe==0.7.0
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.40
tornado==6.2
tzdata==2022.1
tzlocal==4.2
